+++ CHANGES.rst
@@ -5,9 +5,9 @@
 4.4 (unreleased)
 ================
 
-- Add support for Python 3.8.
+- Add support for Python 3.8 and 3.9
 
-- Drop support for Python 3.4.
+- Drop support for Python 3.4 and 3.5.
 
 
 4.3 (2018-10-04)